// src/components/ui/paper-onboarding/types.d.ts
export interface PaperOnboardingMethods {
  /**
   * Navigate to next page.
   */
  next: () => void;
  /**
   * Navigate to previous page.
   */
  previous: () => void;
}
// ...
export interface PaperOnboardingScreenDimensions {
  width: number;
  height: number;
}
// ...
export interface PaperOnboardingProps
  extends